+++ 3DRadSpace/Engine3DRadSpace/Content/ContentManager.cpp
@@ -1,5 +1,4 @@
 #include "ContentManager.hpp"
-#include "ContentManager.h"
 #include "../Logging/Logging.hpp"
 #include "../Graphics/IGraphicsDevice.hpp"
 #include "Assets/Assets.hpp"
@@ -108,39 +107,4 @@ size_t ContentManager::Count() const noexcept
 IGraphicsDevice* ContentManager::GetDevice() const noexcept
 {
 	return _owner->GetService<IGraphicsDevice>({});
-}
-
-E3DRSP_IAsset E3DRSP_ContentManager_Load(E3DRSP_ContentManager content, const E3DRSP_UUID* uuid, const char* path, unsigned* id)
-{
-	if (content == nullptr) return nullptr;
-
-	return static_cast<ContentManager*>(content)->Load(
-		*reinterpret_cast<const Reflection::UUID*>(uuid),
-		std::filesystem::path(path),
-		id
-	);
-}
-
-void E3DRSP_ContentManager_Reload(E3DRSP_ContentManager content, unsigned id)
-{
-	if (content == nullptr) return;
-	static_cast<ContentManager*>(content)->Reload(id);
-}
-
-void E3DRSP_ContentManager_Remove(E3DRSP_ContentManager content, unsigned id)
-{
-	if (content == nullptr) return;
-	static_cast<ContentManager*>(content)->RemoveAsset(id);
-}
-
-void E3DRSP_ContentManager_Clear(E3DRSP_ContentManager content)
-{
-	if (content == nullptr) return;
-	static_cast<ContentManager*>(content)->Clear();
-}
-
-size_t E3DRSP_ContentManager_Count(E3DRSP_ContentManager content)
-{
-	if (content == nullptr) return 0;
-	return static_cast<ContentManager*>(content)->Count();
 }
